move groupById to its own file
