+++ server/src/lib/data/models/group-by-id.ts
@@ -1,0 +1,18 @@
+import type { ById } from "@t/data/utility.types";
+
+type IdFieldUnion<T> = {
+	[K in keyof T]: K extends `${string}_id` ? K : never;
+}[keyof T];
+
+/** Reduces a list of objects that share an *_id field into a Record<ID,
+ * object>. */
+export function groupById<T extends object>(
+	data: T[],
+	idField: IdFieldUnion<T>,
+): ById<T> {
+	return data.reduce((acc, cur) => {
+		const a = cur[idField] as number;
+		acc[a] = cur;
+		return acc;
+	}, {} as ById<T>);
+}
